requirements.txt
backoff==2.2.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Simple-GeoIP==0.2.4
idna==3.7
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
simple-geoip==0.1.1
urllib3==2.2.2
Werkzeug==3.0.3
